tests/requirements.txt
appnope==0.1.3
APScheduler==3.6.3
asttokens==2.0.5
backcall==0.2.0
bleach==5.0.1
certifi==2017.7.27.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
cryptography==37.0.2
cycler==0.10.0
decorator==4.0.11
Django==3.1
django-admin-list-filter-dropdown==1.0.2
django-environ==0.4.5
django-json-widget==1.1.1
django-model-utils==4.2.0
entrypoints==0.2.2
et-xmlfile==1.1.0
executing==0.8.3
future==0.16.0
html5lib==0.999
idna==2.5
ipykernel==4.5.2
ipython==8.1.0
ipython-genutils==0.1.0
jedi==0.18.1
Jinja2==2.9.5
jsonschema==2.5.1
jupyter-client==7.2.0
jupyter-core==4.10.0
MarkupSafe==2.1.1
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
nest-asyncio==1.5.5
networkx==1.11
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.4
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.3.6
pycparser==2.21
Pygments==2.12.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-telegram-bot==13.15
pyzmq==23.0.0
rsa==4.8
simplegeneric==0.8.1
six==1.12.0
sqlparse==0.4.2
stack-data==0.2.0
testpath==0.3
tzdata==2022.1
tzlocal==4.2
vine==5.0.0
wcwidth==0.1.7
webencodings==0.5.1